+++ frontend/src/pages/helpcenter/station/basics/videos.vue
@@ -1,0 +1,17 @@
+/*
+ *     SPDX-License-Identifier: AGPL-3.0-only
+ *
+ *     Copyright (C) RainbowDashLabs and Contributor
+ */
+<script setup lang="ts">
+import VideosHelp from '~/views/helpcenter/basics/VideosHelp.vue'
+
+definePageMeta({
+  layout: 'helpcenter',
+  name: 'help-basics-videos',
+})
+</script>
+
+<template>
+  <VideosHelp />
+</template>